Rename files to avoid collisions
